requirements.txt
asgiref==3.8.1
black==24.4.2
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.0.6
django-js-asset==2.2.0
django-mptt==0.16.0
django-polymorphic==3.1.0
djangorestframework==3.15.1
djangorestframework-recursive==0.1.2
drf-nested-routers==0.94.1
easy-thumbnails==2.8.5
filelock==3.15.4
flake8==7.1.0
identify==2.5.36
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pre-commit==3.7.1
pycodestyle==2.12.0
pyflakes==3.2.0
PyYAML==6.0.1
ruff==0.4.10
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.26.3
